% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pii_table.R
\name{pii_table}
\alias{pii_table}
\title{Creates a summary table of PII for a dataset.}
\usage{
pii_table(df, path)
}
\arguments{
\item{df}{A dataframe with (vector-able) variables.}
}
\description{
This function takes a dataframe and searches each column for presence of 
any text that appears as PII: emails, DOB, SSNs, and phone numbers.
}
\examples{
set.seed(2)
testcase <- data.frame(name = generator::r_full_names(2),  
snn = generator::r_national_identification_numbers(2), 
dob = generator::r_date_of_births(2), 
email = generator::r_email_addresses(2), 
ip = generator::r_ipv4_addresses(2), 
phone = generator::r_phone_numbers(2),  
credit_card = generator::r_credit_card_numbers(2), 
lat = generator::r_latitudes(2), 
lon = generator::r_longitudes(2), 
stringsAsFactors = FALSE)

pii_table(testcase)

}
