% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_addresses.R
\name{search_addresses}
\alias{search_addresses}
\title{Searches vector for full addresses, including zipcode, city, state, street 
name and house number.}
\usage{
search_addresses(vec, output)
}
\arguments{
\item{vec}{A vector input whose contents need to be searched for references
to addresses}

\item{output}{The desired output of function. Defaults to "vector" where T/F 
vector result is returned. The argument "df" will output a table of original 
vector input, T/F vector result, and the matching substring.}
}
\description{
This function takes a vector input and searches presence of any text that 
appears to be a full address. It first identifies whether any string appears 
to be a state name or state abbreviation, then determines whether the 
preceding text appears to be a city in that state. Once a city and state are 
identified, it then looks for strings that appear to be zipcodes or street 
addresses (meaning street name and house number).
}
\details{
The default output is a TRUE/FALSE vector but the function can also
return a dataframe of the original vector input, the TRUE/FALSE result, 
and the substring that matched the address pattern.
}
\examples{

fakeaddresses <- c("820 Nut Swamp Ave.
Toms River, NJ 08753",
                   "982 Clay Street
                   Lenoir, NC 28645",
                   "4 NW. Mayfield Rd.
                   Springfield Gardens, NY 11413",
                   "20 Elmwood Street
                   Raleigh, NC 27603",
                   "196 E. Green Lake Road
                   Birmingham, AL 35209",
                   "73 Beechwood Dr.
                   La Crosse, WI VA DC 54601")

search_addresses(fakeaddresses)

}
