% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_cities.R, R/search_cities_in_states.R
\name{search_cities_in_states}
\alias{search_cities_in_states}
\title{Searches vector for cities (based on whether vector includes state).}
\usage{
search_cities_in_states(vec, output)

search_cities_in_states(vec, output)
}
\arguments{
\item{vec}{A vector input whose contents need to be searched for references
to city names}

\item{output}{The desired output of function. Defaults to "vector" where T/F 
vector result is returned. The argument "dt" will output a table of original 
vector input, T/F vector result, and the matching substring.}
}
\description{
This function takes a vector input and searches presence of #' any text that 
appears to be a zipcode by looking for sequences of numbers. Specifically,
it first identifies whether any string appears to be a state name or state 
abbreviation, then determines whether the preceding text appears to be a city 
in that state.

This function takes a vector input and searches presence of #' any text that 
appears to be a zipcode by looking for sequences of numbers. Specifically,
it first identifies whether any string appears to be a state name or state 
abbreviation, then determines whether the preceding text appears to be a city 
in that state.
}
\details{
The default output is a TRUE/FALSE vector but the function can also
return a dataframe of the original vector input, the TRUE/FALSE result, 
and the substring that matched the zipcode pattern.

The default output is a TRUE/FALSE vector but the function can also
return a dataframe of the original vector input, the TRUE/FALSE result, 
and the substring that matched the city/state pattern.
}
\examples{

fakeaddresses <- c("820 Nut Swamp Ave.
Toms River, NJ 08753",
                   "982 Clay Street
                   Lenoir, NC 28645",
                   "4 NW. Mayfield Rd.
                   Springfield Gardens, NY 11413",
                   "20 Elmwood Street
                   Raleigh, NC 27603",
                   "196 E. Green Lake Road
                   Birmingham, AL 35209",
                   "73 Beechwood Dr.
                   La Crosse, WI VA DC 54601")

search_cities_in_states(testcase)


fakeaddresses <- c("820 Nut Swamp Ave.
Toms River, NJ 08753",
                   "982 Clay Street
                   Lenoir, NC 28645",
                   "4 NW. Mayfield Rd.
                   Springfield Gardens, NY 11413",
                   "20 Elmwood Street
                   Raleigh, NC 27603",
                   "196 E. Green Lake Road
                   Birmingham, AL 35209",
                   "73 Beechwood Dr.
                   La Crosse, WI VA DC 54601")

search_cities_in_states(fakeaddresses)

}
