% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_state.R
\name{search_state}
\alias{search_state}
\title{Searches vector for states by full name and two-letter abbreviation.}
\usage{
search_state(vec, output)
}
\arguments{
\item{vec}{A vector input whose contents need to be searched for references
to states.}

\item{output}{The desired output of function. Defaults to "vector" where T/F 
vector result is returned. The argument "dt" will output a table of original 
vector input, T/F vector result, and the matching substring.}
}
\description{
This function takes a vector input and searches presence of 
any text that matches states' full names or 2 letter abbreviation.
}
\details{
The default output is a TRUE/FALSE vector but the function can also
return a dataframe of the original vector input, the TRUE/FALSE result, 
and the substring that matched the state name pattern.
}
\examples{

fakeaddresses <- c("820 Nut Swamp Ave.
Toms River, NJ 08753",
                   "982 Clay Street
                   Lenoir, NC 28645",
                   "4 NW. Mayfield Rd.
                   Springfield Gardens, NY 11413",
                   "20 Elmwood Street
                   Raleigh, NC 27603",
                   "196 E. Green Lake Road
                   Birmingham, AL 35209",
                   "73 Beechwood Dr.
                   La Crosse, WI VA DC 54601")

search_state(fakeaddresses)

}
