% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_zipcode.R
\name{search_zipcode}
\alias{search_zipcode}
\title{Searches vector for zipcodes.}
\usage{
search_zipcode(vec, output)
}
\arguments{
\item{vec}{A vector input whose contents need to be searched for references
to zipcodes}

\item{output}{The desired output of function. Defaults to "vector" where T/F 
vector result is returned. The argument "df" will output a table of original 
vector input, T/F vector result, and the matching substring.}
}
\description{
This function takes a vector input and searches presence of any text that 
appears to be a zipcode by looking for sequences of numbers. Specifically,
it looks for 5 digit sequences, 9 digit sequences, and 5 digit sequences 
followed by 4 digit sequences with some kind of separator.
The default output is a TRUE/FALSE vector but the function can also
return a dataframe of the original vector input, the TRUE/FALSE result, 
and the substring that matched the zipcode pattern.
}
\examples{

fakeaddresses <- c("820 Nut Swamp Ave.
Toms River, NJ 08753",
                   "982 Clay Street
                   Lenoir, NC 28645",
                   "4 NW. Mayfield Rd.
                   Springfield Gardens, NY 11413",
                   "20 Elmwood Street
                   Raleigh, NC 27603",
                   "196 E. Green Lake Road
                   Birmingham, AL 35209",
                   "73 Beechwood Dr.
                   La Crosse, WI VA DC 54601")

search_zipcode(fakeaddresses)

}
