% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pii_geo_table.R
\name{pii_geo_table}
\alias{pii_geo_table}
\title{Creates a summary table of PII geography for a dataset.}
\usage{
pii_geo_table(df, path, writeout)
}
\arguments{
\item{df}{A dataframe with (vector-able) variables.}
}
\description{
This function takes a dataframe and searches each column for presence of 
any text that appears as PII geography: street address, city/states, zipcodes.
}
\examples{
fakeaddresses <- c("820 Nut Swamp Ave.
Toms River, NJ 08753",
                   "982 Clay Street
                   Lenoir, NC 28645",
                   "4 NW. Mayfield Rd.
                   Springfield Gardens, NY 11413",
                   "20 Elmwood Street
                   Raleigh, NC 27603",
                   "196 E. Green Lake Road
                   Birmingham, AL 35209",
                   "73 Beechwood Dr.
                   La Crosse, WI VA DC 54601")

pii_geo_table(fakeaddresses)

}
